RMCALL: pass and retrieve DS correctly
If DS which is passed is not zero (in this implementation), further code
does not acquire correct values because it directly depend on DS, therefore
DS must be prepared in the moment when all data related moves are done.

On return from interrupt we first set DS, so that we can access correct
data memory. Setting DS involve first storing returned DS on stack to
be able to put it later to returned values and loading 0 to DS (now both
the code and the data are in first 64kB of memory).
Register esi for storing all registers values was replaced by eax, just
because we already have eax stored on the stack (we are sing ax to set DS).