RMCALL: use 2B registers in .data16 context to address memory instead of 4B registers
It generates shorter instructions, there is no need for prefix 0x67 which
changes default 2B address operand to 4B operand in .data16 context.
There is counted correct DS:BX pointer for memory in/outs.

Jump to real mode must be addressed with 4B register, because the real mode
like data selectors are going to be kept 0 based and structure holding data
for jump could be wherever in the first 1MB, when the rest of the code gets
there.

bx must be used instead of ax after int call, because ax is not usable in
such an instruction as base register.